+++ packages/core/src/interceptor/types.ts
@@ -1,15 +1,25 @@
+/*
+ * © Copyright 2022 HP Development Company, L.P.
+ * SPDX-License-Identifier: MIT
+ */
+
+export interface InterceptorBagDetails {
+	Context?: unknown;
+	State?: unknown;
+}
+
 /**
  * Interceptor bag object
  *
  * @typeParam Context - The type of the context that will be injected
  * @typeParam State - The type of the optional state that can be used to propagate state between interceptors
  */
-export type InterceptorBag<Context = unknown, State = unknown> = {
+export type InterceptorBag<IBD extends InterceptorBagDetails = InterceptorBagDetails, AdditionalProps = {}> = {
 	module: string;
 	handlerArgs: unknown[];
-	context?: Context;
-	state?: State;
-};
+	context?: IBD['Context'];
+	state?: IBD['State'];
+} & AdditionalProps;
 
 /**
  * Interceptor function
@@ -17,10 +27,10 @@ export type InterceptorBag<Context = unknown, State = unknown> = {
  * @typeParam Context - The type of the context that will be injected
  * @typeParam State - The type of the optional state that can be used to propagate state between interceptors
  */
-export type Interceptor<Context = unknown, State = unknown> = (
+export type Interceptor<IBD extends InterceptorBagDetails = InterceptorBagDetails, AdditionalProps = {}> = (
 	// eslint-disable-next-line no-use-before-define
-	next: InterceptorNext<Context, State>,
-	interceptorBag?: InterceptorBag<Context, State>
+	next: InterceptorNext<IBD, AdditionalProps>,
+	interceptorBag?: InterceptorBag<IBD, AdditionalProps>
 ) => any;
 
 /**
@@ -29,4 +39,7 @@ export type Interceptor<Context = unknown, State = unknown> = (
  * @typeParam Context - The type of the context that will be injected
  * @typeParam State - The type of the optional state that can be used to propagate state between interceptors
  */
-export type InterceptorNext<Context = unknown, State = unknown> = () => ReturnType<Interceptor<Context, State>>;
+export type InterceptorNext<
+	IBD extends InterceptorBagDetails = InterceptorBagDetails,
+	AdditionalProps = {}
+> = () => ReturnType<Interceptor<IBD, AdditionalProps>>;