+++ packages/health-checks/test/unit/HealthChecksModule.test.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-import { App } from '@davinci/core';
+import { App, mapObject } from '@davinci/core';
 import { FastifyHttpServer } from '@davinci/http-server-fastify';
 import { createSandbox } from 'sinon';
 import { healthCheck, HealthChecksModule } from '../../src';
@@ -12,6 +12,9 @@ import { expect } from '../support/chai';
 const sinon = createSandbox();
 
 describe('HealthChecksModule', () => {
+	let app: App;
+	afterEach(() => app.shutdown());
+
 	describe('onInit', () => {
 		it('should inspect controllers and initialize correctly', async () => {
 			class MyController {
@@ -21,7 +24,7 @@ describe('HealthChecksModule', () => {
 				@healthCheck('liveness')
 				onLivenessCheck() {}
 			}
-			const app = new App();
+			app = new App({ logger: { level: 'silent' } });
 			app.registerController(MyController);
 			const fastifyHttpServer = new FastifyHttpServer();
 			const healthChecksModule = new HealthChecksModule({
@@ -42,12 +45,40 @@ describe('HealthChecksModule', () => {
 			expect(onInitResult).to.haveOwnProperty('/checks/readiness').to.be.a('function');
 		});
 
+		it('should support multiple decorators for the same method', async () => {
+			class MyController {
+				@healthCheck('readiness')
+				@healthCheck('liveness')
+				check() {}
+			}
+			const checkSpy = sinon.spy(MyController.prototype, 'check');
+			app = new App({ logger: { level: 'silent' } });
+			app.registerController(MyController);
+			const fastifyHttpServer = new FastifyHttpServer();
+			const healthChecksModule = new HealthChecksModule({
+				healthChecks: [
+					{ name: 'liveness', endpoint: '/checks/liveness' },
+					{ name: 'readiness', endpoint: '/checks/readiness' }
+				]
+			});
+			await app.registerModule(fastifyHttpServer);
+			await app.registerModule(healthChecksModule);
+			const onInitSpy = sinon.spy(healthChecksModule, 'onInit');
+
+			await app.init();
+
+			expect(onInitSpy.called).to.be.true;
+			const onInitResult = await onInitSpy.getCall(0).returnValue;
+			await mapObject(onInitResult, checkFn => typeof checkFn === 'function' && checkFn(null));
+			expect(checkSpy.callCount).to.be.equal(2);
+		});
+
 		it('should fail if trying to register a health check not listed in the configuration', async () => {
 			class MyController {
 				@healthCheck('readiness')
 				onReadinessCheck() {}
 			}
-			const app = new App();
+			app = new App();
 			sinon.stub(app, 'getModuleById');
 			app.registerController(MyController);
 			const healthChecksModule = new HealthChecksModule({