+++ packages/health-checks/test/unit/decorators.test.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: MIT
  */
 
-import { reflect } from '@davinci/reflector';
+import { DecoratorId, reflect } from '@davinci/reflector';
 import { healthCheck } from '../../src';
 import { expect } from '../support/chai';
 
@@ -29,6 +29,46 @@ describe('decorators', () => {
 						decorators: [
 							{
 								healthCheckName: 'readiness'
+							}
+						]
+					}
+				],
+				properties: [],
+				ctor: {
+					kind: 'Constructor',
+					name: 'constructor',
+					parameters: []
+				},
+				typeClassification: 'Class'
+			});
+		});
+
+		it('should support multiple decorators', () => {
+			class CustomerController {
+				@healthCheck('readiness')
+				@healthCheck('liveness')
+				check() {}
+			}
+
+			const reflection = reflect(CustomerController);
+
+			expect(reflection).to.containSubset({
+				kind: 'Class',
+				name: 'CustomerController',
+				decorators: [],
+				methods: [
+					{
+						kind: 'Method',
+						name: 'check',
+						parameters: [],
+						decorators: [
+							{
+								[DecoratorId]: 'health-check.method',
+								healthCheckName: 'readiness'
+							},
+							{
+								[DecoratorId]: 'health-check.method',
+								healthCheckName: 'liveness'
 							}
 						]
 					}