+++ packages/openapi/src/index.ts
@@ -4,3 +4,22 @@
  */
 
 export * from './OpenAPIModule';
+
+declare module '@davinci/http-server' {
+	interface MethodDecoratorOptions {
+		/**
+		 * if set to 'true' hides the endpoint from the
+		 * generated OpenAPI document
+		 *
+		 * @defaultValue false
+		 */
+		hidden?: boolean;
+	}
+
+	interface ControllerDecoratorOptions {
+		/**
+		 * Defines the OpenAPI tags for the endpoint
+		 */
+		tags?: Array<string>;
+	}
+}