+++ website/src/pages/index.tsx
@@ -1,0 +1,40 @@
+/*
+ * © Copyright 2022 HP Development Company, L.P.
+ * SPDX-License-Identifier: MIT
+ */
+
+import React from 'react';
+import clsx from 'clsx';
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+import Layout from '@theme/Layout';
+import HomepageFeatures from '@site/src/components/HomepageFeatures';
+import Logo from '@site/static/img/logo.svg';
+
+import styles from './index.module.css';
+
+function HomepageHeader() {
+	const { siteConfig } = useDocusaurusContext();
+	return (
+		<header className={clsx('hero hero--primary', styles.heroBanner)}>
+			<div className="container">
+				<Logo />
+				<h1 className="hero__title">{siteConfig.title}</h1>
+				<p className="hero__subtitle">{siteConfig.tagline}</p>
+			</div>
+		</header>
+	);
+}
+
+export default function Home(): JSX.Element {
+	const { siteConfig } = useDocusaurusContext();
+	return (
+		<Layout title={siteConfig.title} description="Description will go into a meta tag in <head />">
+			<div className={styles.homeWrapper}>
+				<HomepageHeader />
+				<main>
+					<HomepageFeatures />
+				</main>
+			</div>
+		</Layout>
+	);
+}