+++ packages/mongoose/CHANGELOG.md
@@ -3,6 +3,37 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-next.12](https://github.com/HPInc/davinci/compare/@davinci/mongoose@2.0.0-next.11...@davinci/mongoose@2.0.0-next.12) (2022-11-03)
+
+
+### Bug Fixes
+
+* **mongoose:** renamed to davinciContext, changed generateSchema behaviour ([#146](https://github.com/HPInc/davinci/issues/146)) ([90f60dd](https://github.com/HPInc/davinci/commit/90f60ddc292d1f2bac8d5bd14082fbcfad1a0bde))
+
+
+
+
+
+# [2.0.0-next.11](https://github.com/HPInc/davinci/compare/@davinci/mongoose@2.0.0-next.10...@davinci/mongoose@2.0.0-next.11) (2022-11-03)
+
+**Note:** Version bump only for package @davinci/mongoose
+
+
+
+
+
+# [1.1.0](https://github.com/HPInc/davinci/compare/@davinci/mongoose@0.19.0...@davinci/mongoose@1.1.0) (2022-10-05)
+
+
+### Features
+
+* **mongoose:** upgraded mongoose dependency from v5 to v6 and renamed context to davinciContext to avoid type collisions with mongoose context ([4c91957](https://github.com/HPInc/davinci/commit/4c919577e9fb17aa216bed24f60e99778ba17d7f))
+
+
+
+
+
+
 # [0.18.0](https://github.com/HPInc/davinci/compare/@davinci/mongoose@0.17.13...@davinci/mongoose@0.18.0) (2022-06-15)